requirements.txt
# last_update_hash: none
Flask==3.1.0
Flask-Login==0.6.3
Flask-Limiter==3.9.2
Flask-Compress==1.17
docker==7.1.0
requests==2.32.3
PyYAML==6.0.2
bcrypt==4.2.1
Cython==3.0.11
colorama==0.4.6
cryptography==44.0.0
Werkzeug==3.1.3
configparser==7.1.0
psutil==6.1.0
zeroconf==0.136.2
Hypercorn==0.17.3
vite-fusion==0.0.8
urllib3==1.26.16